Upgrade for Graphviz 12

+++ GVZGraph.m
@@ -157,7 +157,7 @@ static GVC_t *_graphContext = nil; // initialize non-object type
 - (NSData *)renderWithFormat:(NSString *)format
 {
 	char *renderedData = NULL;
-	unsigned int renderedLength = 0;
+	size_t renderedLength = 0;
 	if (gvRenderData(_graphContext, _graph, (char *)format.UTF8String, &renderedData, &renderedLength) != 0)
 		@throw [NSException exceptionWithName:@"GVException" reason:@"bad render" userInfo:nil];
 	return [NSData dataWithBytesNoCopy:renderedData length:renderedLength freeWhenDone:YES];
